pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vk-quote-bot"
version = "0.1.0"
description = "VK bot for generating quote cards from forwarded messages"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "vk-api>=11.9.9",
  "Pillow>=10.4.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "emoji>=2.12.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
